+++ src/main/java/com/pm/projectmanager/domain/section/SectionService.java
@@ -10,6 +10,9 @@ import org.springframework.stereotype.Service;
 import com.pm.projectmanager.common.response.ResponseExceptionEnum;
 import com.pm.projectmanager.domain.authority.AuthorityRepository;
 import com.pm.projectmanager.domain.authority.AuthorityService;
+import com.pm.projectmanager.domain.card.Card;
+import com.pm.projectmanager.domain.card.CardRepository;
+import com.pm.projectmanager.domain.comment.CommentRepository;
 import com.pm.projectmanager.domain.project.Project;
 import com.pm.projectmanager.domain.project.ProjectRepository;
 import com.pm.projectmanager.domain.section.dto.SectionCreateDto;
@@ -18,8 +21,10 @@ import com.pm.projectmanager.domain.section.dto.SectionUpdateDto;
 import com.pm.projectmanager.exception.AuthorityNullException;
 import com.pm.projectmanager.exception.ProjectNullException;
 import com.pm.projectmanager.exception.SectionException;
+import com.pm.projectmanager.exception.UserRoleException;
 import com.pm.projectmanager.security.UserDetailsImpl;
 
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -30,6 +35,8 @@ public class SectionService {
 	private final ProjectRepository projectRepository;
 	private final AuthorityRepository authorityRepository;
 	private final AuthorityService authorityService;
+	private final CommentRepository commentRepository;
+	private final CardRepository cardRepository;
 
 	public void create(Long projectId, SectionCreateDto requestDto, UserDetailsImpl userDetails) {
 
@@ -72,11 +79,11 @@ public class SectionService {
 			.collect(Collectors.toList());
 	}
 
+	@Transactional
 	public void update(Long projectId, Long sectionId, SectionUpdateDto requestDto, UserDetailsImpl userDetails) {
-		authorityService.adminCheck(projectId, userDetails.getUser().getId());
-
-		authorityCheck(projectId, userDetails);
-
+		if (!authorityService.adminCheck(projectId, userDetails.getUser().getId())) {
+			throw new UserRoleException(ResponseExceptionEnum.ADMIN_ROLE_REQUIRED);
+		}
 		Section section = sectionRepository.findById(sectionId)
 			.orElseThrow(() -> new SectionException(ResponseExceptionEnum.SECTION_NOT_FOUND));
 
@@ -84,14 +91,19 @@ public class SectionService {
 		sectionRepository.save(section);
 	}
 
+	@Transactional
 	public void delete(Long projectId, Long sectionId, UserDetailsImpl userDetails) {
-		authorityService.adminCheck(projectId, userDetails.getUser().getId());
-
-		authorityCheck(projectId, userDetails);
-
+		if (!authorityService.adminCheck(projectId, userDetails.getUser().getId())) {
+			throw new UserRoleException(ResponseExceptionEnum.ADMIN_ROLE_REQUIRED);
+		}
 		Section section = sectionRepository.findById(sectionId)
 			.orElseThrow(() -> new SectionException(ResponseExceptionEnum.SECTION_NOT_FOUND));
 
+		List<Card> cards = cardRepository.findAllBySectionId(sectionId);
+		for (Card card : cards) {
+			commentRepository.deleteAllByCardId(card.getId());
+		}
+		cardRepository.deleteAll(cards);
 		sectionRepository.delete(section);
 	}
 
@@ -100,4 +112,5 @@ public class SectionService {
 			throw new AuthorityNullException(ResponseExceptionEnum.AUTHORITY_NULL_EXCEPTION);
 		}
 	}
+
 }
